Code
This project aims to apply Machine Learning ML, such as the SVM and NB algorithms to analyze the cryptocurrency users’ sentiments. Furthermore, Deep Learning DL models such as LSTM will forecast the cryptocurrencies’ prices. Multi datasets will be used to achieve the purpose of this research; The first dataset from the Twitter social media platform depends on specific hashtags and keywords. The other datasets will be used to determine the price of cryptocurrencies. In addition, this research studies the correlation between cryptocurrency users’ sentiments and price volatility in the cryptocurrency market and compares the results with previous studies in the same field.